Moved StackPostfixNotation.java from the Others section to the Stack section
